+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
 
-## 1.0.1 - 2023-08-20
-* Split Functions Into a Different File (s3_operations.go)
-* Recursive File/Folder Deletion
+## 1.0.2 - 2023-08-21
+* List Buckets and Objects
+* Downloading Files